26.1.Validation Auto-Mapping
Head over to /admin/article and log in as an admin user: admin1@thespacebar.com password engage. Use this unchecked admin power to go to /admin/article and click to create a new article.
I love the new "secrets" feature... but what I'm about to show you might be my second favorite new thing. It actually comes from Symfony 4.3 but was improved in 4.4. It's called: validation auto-mapping... and it's one more step towards robots doing my programming for me.
Start by going into templates/article_admin/_form.html.twig. This is the form that renders the article admin page. To help us play with validation, on the button, add a formnovalidate attribute.
Thanks to that, after you refresh, HTML5 validation is disabled and we can submit the entire form blank to see... our server-side validation errors. These come from the annotations on the Article class, like @Assert\NotBlank above $title. So it's no surprise that if we remove the @Assert\NotBlank annotation... I'll move it as a comment below the property. That's as good as deleting it. And then re-submit the blank form... the validation error is gone from that field.