26.2.Validation Auto-Mapping{@Assert\EnableAutoMapping: composer show symfony/property-info}
@Assert\EnableAutoMapping
Ready for the magic? Go back to Article and, on top of the class, add @Assert\EnableAutoMapping().
As soon as we do that, we can refresh to see... Kidding! We refresh to see... the validation error is back for the title field!

  This value should not be null

Yep! A @NotNull constraint was automatically added to the property! How the heck did that work? The system - validation auto-mapping - automatically adds sensible validation constraints based off of your Doctrine metadata. The Doctrine Column annotation has a nullable option and its default value is nullable=false.
In other words, the title column is required in the database! And so, a constraint is added to make it required on the form.
Auto-mapping can also add constraints based solely on how your code is written... I'll show you an example of that in a few minutes. Oh, and by the way, to get the most out of this feature, make sure you have the symfony/property-info component installed.

composer show symfony/property-info
If that package doesn't come up, install it to allow the feature to get as much info as possible.