26.3.Validation Auto-Mapping{Auto-Mapping is Smart}
Auto-Mapping is Smart
Let's play with this a bit more, like change this to nullable=true.
This means that the column should now be optional in the database. What happens when we submit the form now? The validation error is gone: the NotNull constraint was not added.
Oh, but it gets even cooler than this. Remove the @ORM\Column entirely - we'll pretend like this property isn't even being saved to the database. I also need to remove this @Gedmo\Slug annotation to avoid an error.
What do you think will happen now? Well think about it: the auto-mapping system won't be able to ask Doctrine if this property is required or not... so my guess is that it won't add any constraints. Try it! Refresh!
Duh, duh, duh! The NotNull validation constraint is back! Whaaaaat? The Doctrine metadata is just one source of info for auto-mapping: it can also look directly at your code. In this case, Symfony looks for a setter method. Search for setTitle():

 326 lines  src/Entity/Article.php
class Article
{
    public function setTitle(string $title): self
    {
        $this->title = $title;
        return $this;
    }
}
Ah yes, the  argument is type-hinted with string. And because that type-hint does not allow null, it assumes that  must be required and adds the validation constraint.