26.5.Validation Auto-Mapping{Avoiding Duplicate Constraints; Automatic Length Annotation; Disabling Auto-Mapping when it Doesn't Make Sense; Configuring Auto-Mapping Globally}
Avoiding Duplicate Constraints
Let's put everything back to where it was in the beginning. What I love about this feature is that... it's just so smart! It accuarely reflects what your code is already communicating.
And even if I add back my @Assert\NotBlank annotation. And refresh... check it out. We don't get 2 errors! The auto-mapping system is smart enough to realize that, because I added a NotBlank annotation constraint to this property, it should not also add the NotNull constraint: that would basically be duplication and the user would see two errors. Like I said, it's smart.

Automatic Length Annotation
And it's not all about the NotNull constraint. The length of this column in the database is 255 - that's the default for a string type. Let's type a super-creative title over and over and over and over again... until we know that we're above that limit. Submit and... awesome:

  This value is too long. It should have 255 characters or less.

Behind-the-scenes, auto-mapping also added an @Length annotation to limit this field to the column size. Say goodbye to accidentally allowing large input... that then gets truncated in the database.

Disabling Auto-Mapping when it Doesn't Make Sense
As cool as this feature is, automatic functionality will never work in all cases. And that's fine for two reasons. First, it's your choice to opt-into this feature by adding the @EnableAutoMapping annotation:

 326 lines  src/Entity/Article.php
use Symfony\Component\Validator\Constraints as Assert;
/**
 * @Assert\EnableAutoMapping()
 */
class Article
{
}
And second, you can disable it on a field-by-field basis.
A great example of when this feature can be a problem is in the User class. Imagine we added @EnableAutoMapping here and created a registration form bound to this class. Well... that's going to be a problem because it will add a NotNull constraint to the $password field! And we don't want that!

 285 lines  src/Entity/User.php
class User implements UserInterface
{
    /**
     * @ORM\Column(type="string", length=255)
     */
    private $password;
}
In a typical registration form - like the one that the make:registration-form command creates - the $password property is set to its hashed value only after the form is submitted & validated. Basically, this is not a field the user sets directly and having the NotNull constraint causes a validation error on submit.
How do you solve this? You could disable auto-mapping for the whole class. Or, you could disable it for the  property only by adding @Assert\DisableAutoMapping:

// src/Entity/User.php

class User implements UserInterface
{
    // ...
    /**
     * @ORM\Column(type="string", length=255)
     * @Assert\DisableAutomapping()
     */
    private $password;
    // ...
}
This is the one ugly case for this feature, but it's easy to fix.

Configuring Auto-Mapping Globally
Oh, and one more thing! You can control the feature a bit in config/packages/validator.yaml. By default, you need to enable auto-mapping on a class-by-class basis by adding the @Assert\EnableAutoMapping annotation:

 326 lines  src/Entity/Article.php
use Symfony\Component\Validator\Constraints as Assert;
/**
 * @Assert\EnableAutoMapping()
 */
class Article
{
}
But, you can also automatically enable it for specific namespaces:
9 lines  config/packages/validator.yaml
framework:
    validation:
        # Enables validator auto-mapping support.
        # For instance, basic validation constraints will be inferred from Doctrine's metadata.
        #auto_mapping:
        #    App\Entity\: []
If we uncommented this App\Entity line, every entity would get auto-mapped validation without needing the extra annotation. I like being a bit more explicit - but it's your call.
Next, ready to talk about something super geeky? No, not Star Trek, but that would awesome. This is probably even better: let's chat about password hashing algorithms. Trust me, it's actually pretty neat stuff. Specifically, I want to talk about safely upgrading hashed passwords in your database to stay up-to-date with security best-practices.